requirements.txt
Jinja2==3.1.5
python-dotenv==1.0.1
requests==2.32.3

Flask~=3.1.0
SQLAlchemy~=2.0.38
typing~=3.7.4.3
pip~=25.0.1
attrs~=25.1.0
wheel~=0.45.1
pillow~=11.0.0
docutils~=0.21.2
tornado~=6.4.2
setuptools~=75.8.0
packaging~=24.1
pyparsing~=3.2.0
typing_extensions~=4.12.2
matplotlib~=3.10.0
statistics~=1.0.3.5